clean up changes
