+++ client/src/components/quiz/QuestionCard.jsx
@@ -3,7 +3,6 @@ import {
   Card,
   CardContent,
   CardActions,
-  Button,
   Typography,
   FormControl,
   FormControlLabel,
